% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestor_seq.R
\name{adjust_ally}
\alias{adjust_ally}
\title{adjust_ally}
\usage{
adjust_ally(tree, node, sub = FALSE, seq_colname = "mol_seq")
}
\arguments{
\item{tree}{ggtree object}

\item{node}{internal node in tree}

\item{sub}{logical value.}

\item{seq_colname}{the colname of MSA on tree$data}
}
\value{
tree
}
\description{
adjust the tree branch position after assigning ancestor node
}
\author{
Lang Zhou
}
