% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dms.R
\name{assign_dms}
\alias{assign_dms}
\title{assign_dms}
\usage{
assign_dms(x, dms)
}
\arguments{
\item{x}{data frame from tidy_msa()}

\item{dms}{dms data frame}
}
\value{
tree
}
\description{
assign dms value to alignments.
}
\author{
Lang Zhou
}
