% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{assign_mds}
\alias{assign_mds}
\title{assign_mds}
\usage{
assign_mds(x, mds)
}
\arguments{
\item{x}{data frame from tidy_msa()}

\item{mds}{MDS data frame}
}
\value{
tree
}
\description{
assign MDS value to alignments.
}
\author{
Lang Zhou
}
