% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{available_fonts}
\alias{available_fonts}
\title{List Font Families currently available}
\usage{
available_fonts()
}
\value{
A character vector of available font family names
}
\description{
This function lists font families currently available that can be used by 'ggmsa'
}
\examples{
available_fonts()
}
\author{
Lang Zhou
}
