% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{available_msa}
\alias{available_msa}
\title{List MSA objects currently available}
\usage{
available_msa()
}
\value{
A character vector of available objects
}
\description{
This function lists MSA objects currently available that can be used by 'ggmsa'
}
\examples{
available_msa()
}
\author{
Lang Zhou
}
