% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustal.R
\name{Clustal}
\alias{Clustal}
\title{The color scheme of Culstal}
\usage{
Clustal(color, alndf)
}
\arguments{
\item{color}{a data frame created manually for Clutal color assignment.}

\item{alndf}{data frame of alignment}
}
\description{
This function is a algorithm to assign color for Multiple Sequence .
}
\keyword{culstal}
