% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_else.R
\name{color_scheme}
\alias{color_scheme}
\title{Assigning colors to sequence alignment.}
\usage{
color_scheme(y, color)
}
\arguments{
\item{y}{sequence alignment with data frame, generated by tidy_msa().}

\item{color}{a Color scheme. One of 'Clustal', 'Chemistry_AA', 'Shapely_AA', 'Zappo_AA', 'Taylor_AA', 'LETTER', 'CN6', 'Chemistry_NT', 'Shapely_NT', 'Zappo_NT', 'Taylor_NT'. Defaults is 'Chemistry_AA'.}
}
\description{
Assigning colors to sequence alignment.
}
