% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_else.R
\name{color_scheme_}
\alias{color_scheme_}
\title{The color scheme of Culstal}
\usage{
color_scheme_(color)
}
\arguments{
\item{color}{a data frame created manually for color assignment.}
}
\description{
This function is a algorithm to assign color for Multiple Sequence .
}
\keyword{color}
\keyword{scheme}
