% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestor_seq.R
\name{extract_seq}
\alias{extract_seq}
\title{extract_seq}
\usage{
extract_seq(tree_adjust)
}
\arguments{
\item{tree_adjust}{ggtree object}
}
\value{
character
}
\description{
extract ancestor sequence from tree data
}
\author{
Lang Zhou
}
