% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_msa.R
\name{facet_msa}
\alias{facet_msa}
\title{segment MSA}
\usage{
facet_msa(field)
}
\arguments{
\item{field}{a numeric vector of the field size}
}
\description{
The MSA would be plot in a field that you set.
}
\examples{
library(ggplot2) 
f <- system.file("extdata/sample.fasta", package="ggmsa")
# 2 fields
ggmsa(f, end = 120, font = NULL, color="Chemistry_AA") + facet_msa(field = 60)
# 3 fields 
ggmsa(f, end = 120, font = NULL,  color="Chemistry_AA") + facet_msa(field = 40)
}
\author{
Lang Zhou
}
