% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_msa.R
\name{geom_msa}
\alias{geom_msa}
\title{geom_msa}
\usage{
geom_msa(data, font = "helvetical", color = "Clustal",
  char_width = 0.9, seq_logo = FALSE, ...)
}
\arguments{
\item{data}{multiple sequence alignment in tidy data frame, generated by tidy_fasta()}

\item{font}{font families, possible values are 'helvetical', 'times', and 'mono'. Defaults is 'helvetical'. If you specify font = NULL, only the background box will be printed.}

\item{color}{A Color scheme. One of 'Clustal', 'Chemistry_AA', 'Shapely_AA', 'Zappo_AA', 'Taylor_AA', 'Chemistry_NT', 'Shapely_NT', 'Zappo_NT', 'Taylor_NT'. Defaults is 'Clustal'.}

\item{char_width}{characters width. Defaults is 0.9.}

\item{seq_logo}{If TRUE plot sequence logo for nucleotide sequences, if FALSE (the default) don't shouw it.}

\item{...}{additional parameter}
}
\value{
A list
}
\description{
Multiple sequence alignment layer for ggplot2. It creats background frames with/without sequence labels.
}
\author{
guangchuang yu
}
