% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_msaBar.R
\name{geom_msaBar}
\alias{geom_msaBar}
\title{geom_msaBar}
\usage{
geom_msaBar()
}
\value{
A list
}
\description{
Multiple sequence alignment layer for ggplot2. It plot sequence conservation bar.
}
\examples{
#plot multiple sequence alignment and conservation bar.
f <- system.file("extdata/sample.fasta", package="ggmsa")
ggmsa(f, 221, 280, font = NULL, seq_name = TRUE) + geom_msaBar()
}
\author{
Lang Zhou
}
