% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_seed.R
\name{geom_seed}
\alias{geom_seed}
\title{geom_seed}
\usage{
geom_seed(msa, seed, star = F)
}
\arguments{
\item{msa}{a multiple sequence alignment file or object}

\item{seed}{character, like 'GAGGUAG' a miRNA seed sequence}

\item{star}{a logical value indicating whether symbol of stars should be produced}
}
\value{
a ggplot layer
}
\description{
highlight the seed in miRNA sequences(test)
}
\details{
We're trying to develop this function.
}
\examples{
miRNA_sequences <- system.file("extdata/seedSample.fa", package="ggmsa")
ggmsa(miRNA_sequences, font = 'DroidSansMono', color = "Chemistry_NT", none_bg = TRUE) + 
geom_seed(miRNA_sequences, seed = "GAGGUAG") 
ggmsa(miRNA_sequences, font = 'DroidSansMono', color = "Chemistry_NT") + 
geom_seed(miRNA_sequences, seed = "GAGGUAG", star = TRUE)
}
\author{
Lang Zhou
}
