% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{geom_seqlogo}
\alias{geom_seqlogo}
\title{geom_seqlogo}
\usage{
geom_seqlogo(
  msa,
  start = NULL,
  end = NULL,
  font = "helvetical",
  color = "Chemistry_NT"
)
}
\arguments{
\item{msa}{multiple sequence alignment file or
sequence object in DNAStringSet, RNAStringSet, AAStringSet, BStringSet,
DNAMultipleAlignment, RNAMultipleAlignment, AAMultipleAlignment, DNAbin or AAbin}

\item{start}{start position to extract subset of alignment}

\item{end}{end position to extract subset of alignemnt}

\item{font}{font families, possible values are 'helvetical', 'times', and 'mono'. Defaults is 'helvetical'.}

\item{color}{A Color scheme. One of 'Chemistry_NT', 'Shapely_NT', 'Zappo_NT', 'Taylor_NT'. Defaults is 'Chemistry_NT'.}
}
\value{
A list
}
\description{
Multiple sequence alignment layer for ggplot2. It creats sequence logo.
}
\examples{
#plot multiple sequence alignment and sequence logo 
f <- system.file("extdata/LeaderRepeat_All.fa", package="ggmsa")
ggmsa(f,font = NULL,color="Chemistry_NT") + geom_seqlogo(f)
}
\author{
Lang Zhou
}
