% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cons.R
\name{get_consensus}
\alias{get_consensus}
\title{calling the consensus sequence.}
\usage{
get_consensus(tidy, ignore_gaps = FALSE, ref = NULL)
}
\arguments{
\item{tidy}{sequence alignment with data frame, generated by tidy_msa().}

\item{ignore_gaps}{a logical value. When selected TRUE, gaps in column are treated as if that row didn't exist.}

\item{ref}{a character string. Specifying the reference sequence which should be one of input sequences when 'consensus_views' is TRUE.}
}
\description{
calling the consensus sequence.
}
\keyword{get_consensus}
