% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmaf.R
\name{ggmaf}
\alias{ggmaf}
\title{ggmaf}
\usage{
ggmaf(
  data,
  ref,
  block_start = NULL,
  block_end = NULL,
  facet_field = NULL,
  heights = c(0.4, 0.6),
  facet_heights = NULL
)
}
\arguments{
\item{data}{a tidy MAF data frame.You can get it by tidy_maf_df()}

\item{ref}{character, the name of reference genome. 
eg:"hg38.chr1_KI270707v1_random"}

\item{block_start}{a numeric vector(>0). The start block to plot.}

\item{block_end}{a numeric vector(< max block). The end block to plot.}

\item{facet_field}{a numeric vector. The field in a facet panel.}

\item{heights}{two numeric vector.The plot proportion between 
"Genomic location" panel(upon) and "Alignment" panel(down).
Default:c(0.4,0.6)}

\item{facet_heights}{Numeric vectors.The facet proportion.}
}
\value{
ggplot object
}
\description{
plot MAF
}
\author{
Lang Zhou
}
