% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Gram-positive_AKL.fasta}
\alias{Gram-positive_AKL.fasta}
\title{Gram-positive_AKL}
\format{
A MSA fasta with 100 sequences and 36 positions.
}
\source{
\url{http://biovis.net/year/2013/info/redesign-contest}
}
\usage{
system.file("extdata", "Gram-positive_AKL.fasta", package="ggmsa")
}
\description{
Amino acids in the adenylate kinase lid (AKL) domain
from Gram-positive bacteria.
}
\keyword{datasets}
