% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LeaderRepeat_All.fa}
\alias{LeaderRepeat_All.fa}
\title{A sample DNA alignment sequences}
\format{
A MSA fasta
}
\description{
DNA alignment sequences with 24 sequences and 56 positions.
}
\keyword{datasets}
