% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_interactive.R
\name{merge_seq}
\alias{merge_seq}
\title{merge_seq}
\usage{
merge_seq(previous_seq, gap, subsequent_seq, adjust_name = TRUE)
}
\arguments{
\item{previous_seq}{previous MSA}

\item{gap}{gap length}

\item{subsequent_seq}{subsequent MSA}

\item{adjust_name}{logical value. merge seq name or not}
}
\value{
tidy MSA data frame
}
\description{
merge two MSA
}
\author{
Lang Zhou
}
