% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa_data.R
\name{msa2tidy}
\alias{msa2tidy}
\title{This function converts the msa_data to the tidy data.}
\usage{
msa2tidy(msaData)
}
\arguments{
\item{msaData}{sequence alignment data generated by msa_data().}
}
\description{
This function converts the msa_data to the tidy data.
}
