% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa_data.R
\name{msa_data}
\alias{msa_data}
\title{Get the data of sequence logos and colors from MSA.}
\usage{
msa_data(fasta, start = NULL, end = NULL, font = "helvetica_regular",
  color = "Clustal")
}
\arguments{
\item{fasta}{Aligned fasta file.}

\item{start}{Start position to plot,Defaults = NULL.}

\item{end}{End position to plot, Defaults = NULL.}

\item{font}{Character font, Defaults is 'helvetica_regular'.}

\item{color}{A Color scheme. One of 'Clustal', 'Chemistry_AA', 'Shapely_AA', 'Zappo_AA', 'Taylor_AA', 'Chemistry_NT', 'Shapely_NT', 'Zappo_NT', 'Taylor_NT'.Defaults is 'Clustal'.}
}
\value{
A data frame
}
\description{
The fasta is organized into a tidy data set by tidy_data().
Each molecule(amino acid or nucleotide) will be  draw as sequence logo and assigned color according to color schemes. The information about sequence logos and colors is stored in a data frame.
}
\examples{
data <- msa_data(fasta, 20, 120, font = 'helvetica_regular', color = 'Clustal' )
data <- msa_data(fasta, 20, 120, font = 'helvetica_regular', color = 'Chemistry_AA' )
data <- msa_data(fasta, 20, 120, font = 'helvetica_regular', color = 'Chemistry_NT' )
}
\author{
guangchuang yu
}
