% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fasta.R
\name{read_fasta}
\alias{read_fasta}
\title{read_fasta}
\usage{
read_fasta(fasta)
}
\arguments{
\item{fasta}{an XstringSet object frome Biostrings or a filepath}
}
\value{
DNAbin or AAbin object
}
\description{
read sequences
}
\details{
This function supports both NT or AA sequences; It supports multiple input formats such as "DNAStringSet", "BStringSet", "AAStringSet", DNAbin", "AAbin" and a filepath.
}
\author{
Lang Zhou
}
