% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_maf.R
\name{read_maf}
\alias{read_maf}
\title{read_maf}
\usage{
read_maf(multiple_alignment_format)
}
\arguments{
\item{multiple_alignment_format}{a multiple alignment format(MAF) file}
}
\value{
data frame
}
\description{
read 'multiple alignment format'(MAF) file
}
\author{
Lang Zhou
}
