% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-functions.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{readHelix}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{R4RNA}{\code{\link[R4RNA:readStructure]{readHelix}}}
}}

