% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample.fasta}
\alias{sample.fasta}
\title{A sample data used in ggmsa}
\format{
A MSA fasta with 9 sequences and 456 positions.
}
\description{
A dataset containing the alignment sequences of 
the phenylalanine hydroxylase protein (PH4H) 
within nine species
}
\keyword{datasets}
