% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seedSample.fa}
\alias{seedSample.fa}
\title{microRNA data used in ggmsa}
\format{
A MSA fasta with 6 sequences and 22 positions.
}
\source{
\url{https://www.mirbase.org/ftp.shtml}
}
\usage{
system.file("extdata", "seedSample.fa", package = "ggmsa")
}
\description{
Fasta format sequences of mature miRNA sequences 
from miRBase
}
\keyword{datasets}
