% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sequence-link-tree.fasta}
\alias{sequence-link-tree.fasta}
\title{sequence-link-tree}
\format{
A MSA fasta with 28 sequences and 480 positions.
}
\usage{
system.file("extdata", "sequence-link-tree.fasta", package = "ggmsa")
}
\description{
Alignment sequences used to demonstrate circular MSA layout
}
\keyword{datasets}
