% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_interactive.R
\name{simplify_hdata}
\alias{simplify_hdata}
\title{simplify_hdata}
\usage{
simplify_hdata(hdata, sim_msa)
}
\arguments{
\item{hdata}{data from tidy_hdata()}

\item{sim_msa}{MSA data frame}
}
\value{
data frame
}
\description{
reset hdata data position
}
\author{
Lang Zhou
}
