% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cons.R
\name{tidy_color}
\alias{tidy_color}
\title{cleaning the needless sequeces' color according to the consensus sequcence (only used in the consensus views).}
\usage{
tidy_color(y, consensus, disagreement, ref)
}
\arguments{
\item{y}{a data frame, sequence alignment with specified color.}

\item{consensus}{the consensus sequence which can be called by get_consensus().}

\item{disagreement}{a logical value. Displays characters that disagreememt to consensus(excludes ambiguous disagreements).}

\item{ref}{a character string. Specifying the reference sequence which should be one of input sequences when 'consensus_views' is TRUE.}
}
\description{
cleaning the needless sequeces' color according to the consensus sequcence (only used in the consensus views).
}
\keyword{tidy_color}
