% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_interactive.R
\name{tidy_hdata}
\alias{tidy_hdata}
\title{tidy_hdata}
\usage{
tidy_hdata(gap, inter, previous_seq, subsequent_seq)
}
\arguments{
\item{gap}{gap length}

\item{inter}{protein-protein interactive position data}

\item{previous_seq}{previous MSA}

\item{subsequent_seq}{subsequent MSA}
}
\value{
helix data
}
\description{
tidy protein-protein interactive position data
}
\author{
Lang Zhou
}
