% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmaf.R
\name{tidy_maf_df}
\alias{tidy_maf_df}
\title{tidy_maf_df}
\usage{
tidy_maf_df(maf_df, ref)
}
\arguments{
\item{maf_df}{a MAF data frame.You can get it by read_maf()}

\item{ref}{character, the name of reference genome. 
eg:"hg38.chr1_KI270707v1_random"}
}
\value{
data frame
}
\description{
tidy MAF data frame
}
\author{
Lang Zhou
}
