% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TP53_genes.xlsx}
\alias{TP53_genes.xlsx}
\title{genome locus}
\format{
xlsx
}
\usage{
system.file("extdata", "TP53_genes.xlsx", package = "ggmsa")
}
\description{
The local genome map shows the 30000 sites around the TP53 gene.
}
\keyword{datasets}
