% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustal.R
\name{color_Clustal}
\alias{color_Clustal}
\title{A color scheme of Culstal.This function is a algorithm to assign colors for Multiple Sequence.}
\usage{
color_Clustal(y)
}
\arguments{
\item{y}{A data frame, data of a tidy fasta,created by 'tidy_fasta()'.}
}
\description{
A color scheme of Culstal.This function is a algorithm to assign colors for Multiple Sequence.
}
\keyword{clustal}
