% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustal.R
\name{color_Clustal}
\alias{color_Clustal}
\title{A color scheme of Culstal. The algorithm to assign colors for Multiple Sequence.}
\usage{
color_Clustal(y)
}
\arguments{
\item{y}{sequence alignment with data frame, generated by tidy_msa().}
}
\description{
A color scheme of Culstal. The algorithm to assign colors for Multiple Sequence.
}
\keyword{clustal}
