% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_else.R
\name{color_scheme_}
\alias{color_scheme_}
\title{A color scheme.This function assigns colors for Multiple Sequence .}
\usage{
color_scheme_(color)
}
\arguments{
\item{color}{A data frame created manually for color assignment.}

\item{y}{A data frame, data of a tidy fasta, created by 'tidy_fasta()'.}
}
\description{
A color scheme.This function assigns colors for Multiple Sequence .
}
\keyword{color}
\keyword{scheme}
