% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestor_seq.R
\name{extract_seq}
\alias{extract_seq}
\title{extract_seq}
\usage{
extract_seq(tree_adjust, seq_colname = "mol_seq")
}
\arguments{
\item{tree_adjust}{ggtree object}

\item{seq_colname}{the colname of MSA on tree$data}
}
\value{
character
}
\description{
extract ancestor sequence from tree data
}
\author{
Lang Zhou
}
