% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_msa.R
\name{geom_msa}
\alias{geom_msa}
\title{geom_msa}
\usage{
geom_msa(data, font = "helvetica_regular", color = "Clustal", ...)
}
\arguments{
\item{data}{multiple sequence alignment in tidy data frame, generated by tidy_fasta()}

\item{font}{Character font, Defaults is 'helvetica_regular'.}

\item{color}{A Color scheme. One of 'Clustal', 'Chemistry_AA', 'Shapely_AA', 'Zappo_AA', 'Taylor_AA', 'Chemistry_NT', 'Shapely_NT', 'Zappo_NT', 'Taylor_NT'. Defaults is 'Clustal'.}

\item{...}{additional parameter}
}
\value{
A list
}
\description{
Multiple sequence alignment layer for ggplot2. It creats background frames with/without sequence labels.
}
\author{
guangchuang yu
}
