% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_msa.R
\name{geom_msa}
\alias{geom_msa}
\title{geom_msa}
\usage{
geom_msa(data, ...)
}
\arguments{
\item{data}{A data frame generated by mas_data()}

\item{...}{additional parameter}
}
\value{
A list
}
\description{
Multiple sequence alignment layer for ggplot2. It creats background frames with/without sequence labels.
}
\author{
guangchuang yu
}
