% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_seed.R
\name{geom_seed}
\alias{geom_seed}
\title{geom_seed}
\usage{
geom_seed(msa, seed)
}
\arguments{
\item{msa}{a multiple sequence alignment file or object}

\item{seed}{character, like 'GAGGUAG' a miRNA seed sequence}
}
\value{
a ggplot layer
}
\description{
highlight the seed in miRNA sequences(test)
}
\details{
We're trying to develop this function.
}
\examples{
f <- system.file("extdata/seedSample.fa", package="ggmsa")
ggmsa(f, color="Chemistry_NT") + geom_seed(f, seed = "GAGGUAG")
}
\author{
Lang Zhou
}
