% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmsa.R
\name{ggmsa}
\alias{ggmsa}
\title{ggmsa}
\usage{
ggmsa(fasta, start = NULL, end = NULL, font = "helvetica_regular",
  color = c("Clustal", "Chemistry_AA", "Shapely_AA", "Zappo_AA",
  "Taylor_AA", "Chemistry_Nucle", "Shapely_Nucle", "Zappo_Nucle",
  "Taylor_Nucle"))
}
\arguments{
\item{fasta}{aligned fasta file}

\item{start}{start position to plot}

\item{end}{end position to plot}

\item{font}{character font}

\item{color}{5 amino acid color schemes, 4 nucleic acid color schemes. Note: Culstal is an amino acid color scheme}
}
\value{
ggplot object
}
\description{
plot multiple sequence alignment using ggplot2
}
\author{
guangchuang yu
}
