% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmsa.R
\name{ggmsa}
\alias{ggmsa}
\title{ggmsa}
\usage{
ggmsa(fasta, start = NULL, end = NULL, font = "helvetica_regular")
}
\arguments{
\item{fasta}{aligned fasta file}

\item{start}{start position to plot}

\item{end}{end position to plot}

\item{font}{character font}
}
\value{
ggplot object
}
\description{
plot multiple sequence alignment using ggplot2
}
\author{
guangchuang yu
}
