% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmsa.R
\name{ggmsa}
\alias{ggmsa}
\title{ggmsa}
\usage{
ggmsa(
  msa,
  start = NULL,
  end = NULL,
  font = "helvetical",
  color = "Clustal",
  char_width = 0.9
)
}
\arguments{
\item{msa}{Multiple aligned sequence file or object for
representing either nucleotide sequences or peptide sequences.}

\item{start}{Start position to plot, If font=NULL, only the background frame is drawn, and no character.}

\item{end}{End position to plot,If font=NULL, only the background frame is drawn, and no character.}

\item{font}{font families, possible values are 'helvetical', 'times', and 'mono'. Defaults is 'helvetical'. If you specify font = NULL, only the background box will be printed.}

\item{color}{A Color scheme. One of 'Clustal', 'Chemistry_AA', 'Shapely_AA', 'Zappo_AA', 'Taylor_AA', 'Chemistry_NT', 'Shapely_NT', 'Zappo_NT', 'Taylor_NT'. Defaults is 'Clustal'.}

\item{char_width}{characters width. Defaults is 0.9.}
}
\value{
ggplot object
}
\description{
Plot multiple sequence alignment using ggplot2 with multiple color schemes supported.
}
\examples{
#plot multiple sequence alignment
f <- system.file("extdata/sample.fasta", package="ggmsa")
ggmsa(f, 164, 213, color="Chemistry_AA")
}
\author{
guangchuang yu
}
