% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tp53.fa}
\alias{tp53.fa}
\title{TP53 MSA}
\format{
A MSA fasta with 5 sequences and 404 positions.
}
\usage{
system.file("extdata", "tp53.fa", package = "ggmsa")
}
\description{
Alignment sequences of used to show graphical combination
}
\keyword{datasets}
